Cache backend: switch from Django's FileBasedCache to the 3rd-party DiskCache.
This should allow more sensible eviction logic for our cache usage, where we make one-time-use async media keys expire quickly, and make performance keys (e.g. label_details) expire very rarely.
